--- src/components/Icon/index.ts
@@ -1,3 +1,0 @@
-export * from './Arrow'
-export * from './Papular'
-export * from './User'
